H.1.1.0: Fix RabbitMQ ordering
When stopping a RabbitMQ cluster, the only one way to restart it, is to
start the latest online node and then start other nodes.
In the playbook, we stop the service one by one, and then start the last
one which has been stopped to ensure RabbitMQ will start without
failure.

Puppet will take care later of starting other nodes.

Signed-off-by: Emilien Macchi <emilien.macchi@enovance.com>