I.1.0.0: Handle HAproxy service randomly
When running HAproxy with Keepalived in cluster, the service could run
randomly on one of the 3 nodes. Since we make some failovers during the
upgrade process, we can't really ensure where is running the service.
Worst case, HAproxy does not start with an error. But it does not produce
any bug.
Where the VIP it hosted, we will be sure HAproxy will start.