I.1.0.0: enable compute scheduling after upgade process
The scheduling is disabled during the upgrade to avoid any error during
the process. If anyone schedule a VM during the upgrade, the VM could
fail due to databases upgrades.

To avoid this, we disable nova-compute service before the config and
enable it again after the config process, when we are sure nova-compute
is running well.