files/disable-tso: manage virtual routers interfaces
In disable-tso script, add support to manage the interfaces from L3
agent (in q-router namespaces) and allow to disable TSO on it.

This patch adds some functions:
- vm_netlink: disable TSO/GSO/UFO/GRO on a netlink given in argument.
- router_netlink: disable TSO/GSO/UFO/GRO on interfaces within a
  namespace
- ip_monitor: monitor the netlink notifications and run vm_netlink or
  router_netlink on the fly when an interface is created

Also add some comments to explain the expected behavior when using the
functions.

Note: ip_monitor is used "on the fly", while "ip link" command is only
needed to disable TSO after a reboot. The namespaces don't survive to
reboot so we don't need to run router_netlink after a reboot, but only
for regular netlinks.

Co-Authored-By: Sylvain Afchain <sylvain.afchain@enovance.com>