functions: drop update_repositories
moving update_repositories in some functions was a bad idea since the
functions are used to redirect the output to a apt/yum file.
so each time update was run, the ouput was redirected to the file.

let's use update_repositories when we need it, which is the case in this
patch when we want to install ntpstat on Debian.

Signed-off-by: Emilien Macchi <emilien.macchi@enovance.com>