upgrade/I.1.2.0: ensure services are stopped before upgrade
If we don't stop OpenStack services before the upgrade, the Python code
executed can crash and the process is still here and cannot be stopped
by init. We need to kill the process manually.

Let's stop services before the upgrade, and then ensure they are
running after eDeploy job.

Also manage openvswitch-switch service since there is an upgrade in this
release for Debian.