[upgrade] Cleanup Ansible packaging
Because Ansible has been upgraded and the way to package it changed
(/usr/share/ansible does not exist anymore), we have to deal with that
change during the playbook run.
So let's exclude /usr/share/ansible in eDeploy metadata to ensure the
directory survives after the upgrade, and then clean-up with Ansible
itself by the end of the playbook with install_server tag.