I.1.3.0 → J.1.0.0: avoid conflict with mysqlchk
To access MySQL, DB migration scripts use the VIP. If the VIP is on the
first node, the services statues will rely on the node1 haproxy.

The status before the DB migrations is:
- first node is out of the MySQL cluster because the mysqld is stopped
- the other nodes are in the cluster and mysqlchk is on port 9200

If we ask haproxy to check the MySQL servers on port 8200, we will get
all the running nodes out of the cluster, and the migration will
fail.

With this commit, we postpone the mysqlchk port reconfiguration to get
the DB migration.

This commit fixes #138.